+++ truffle/src/com.oracle.truffle.api.strings.test/src/com/oracle/truffle/api/strings/test/ops/TStringByteIndexOfStringTest.java
@@ -98,7 +98,7 @@ public class TStringByteIndexOfStringTest extends TStringTestBase {
         for (int i = 0; i < encodings.length; i++) {
             TruffleString.Encoding encoding = encodings[i];
             byte[] arr = new byte[strA.byteLength(encoding)];
-            strA.switchEncodingUncached(encoding).copyToByteArrayNodeUncached(0, arr, 0, arr.length, encoding);
+            strA.switchEncodingUncached(encoding).copyToByteArrayUncached(0, arr, 0, arr.length, encoding);
             int iFinal = i;
             checkStringVariants(arr, TruffleString.CodeRange.ASCII, true, encoding, codepointsA, null, (a, array, codeRange, isValid, enc, codepoints, byteIndices) -> {
                 Assert.assertEquals(a.byteLength(encoding) - strB.switchEncodingUncached(encoding).byteLength(encoding), node.execute(a, withMask[iFinal], 0, array.length, encoding));